creation of the component how to contribute if I am a restaurant
