+++ src/Home.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 
 import "./Home.css";
 import Whatsapp from "./Componentes/ElementosFijos/Whatsapp";
@@ -6,7 +6,8 @@ import MenuNavegacion from "./Componentes/ElementosFijos/MenuNavegacion";
 import Footer from "./Componentes/ElementosFijos/Footer";
 import ModalCargadoInicial from '../src/Componentes/Modales/ModalInicio';
 
-import Correo from './Componentes/ElementosDinamicos/Correo';
+import ModuloContacto
+ from './Componentes/ElementosDinamicos/ModuloContacto';
 
 function Home() {
 
@@ -20,7 +21,8 @@ function Home() {
     <div className="ContenedorPadre">
       <MenuNavegacion />
       <Whatsapp />
-      <Correo/>
+      <ModuloContacto
+/>
 
       <Footer />
